# Performance improvements on Arm for legacy and k-quants