--- a/llamafile/iqk_mul_mat_arm82.cpp
+++ b/llamafile/iqk_mul_mat_arm82.cpp
@@ -1,0 +1,5 @@
+#ifdef __aarch64__
+#define iqk_mul_mat iqk_mul_mat_arm82
+#define iqk_mul_mat_moe iqk_mul_mat_moe_arm82
+#include "iqk_mul_mat.inc"
+#endif // __aarch64__