@ikawrakow 2024-05-27
I forgot to add a `Q8_0` implementation (required because of the reordering of the quantized activations), so converting to draft until I add it.