@jart 2024-05-29
Here's the improvements on my Mac Studio. Enormous gains for `Q5_K_M`, `Q6_K`, and `Q5_0`!! I'm actually very pleased that you're optimizing the legacy quants too, due to weird new models like [IBM Granite 34b](https://huggingface.co/Mozilla/granite-34b-code-instruct-llamafile).

| cpu_info | model_filename | size | test | t/s before | t/s after | t/s speedup |
| ---: | ---: | ---: | ---: | ---: | ---: | ---: |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q8_0 | 1.09 GiB | pp512 | 693.92 | 883.96 | 1.27x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q8_0 | 1.09 GiB | tg16 | 70.39 | 103.10 | 1.46x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q6_K | 860.86 MiB | pp512 | 222.32 | 617.74 | 2.78x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q6_K | 860.86 MiB | tg16 | 96.01 | 96.93 | 1.01x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q5_K_M | 745.11 MiB | pp512 | 244.09 | 658.62 | 2.70x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q5_K_M | 745.11 MiB | tg16 | 93.74 | 103.06 | 1.10x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q5_0 | 729.84 MiB | pp512 | 245.62 | 809.91 | 3.30x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q5_0 | 729.84 MiB | tg16 | 96.11 | 106.78 | 1.11x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q4_0 | 606.53 MiB | pp512 | 625.47 | 943.14 | 1.51x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q4_0 | 606.53 MiB | tg16 | 129.34 | 124.60 | 0.96x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q2_K | 411.41 MiB | pp512 | 249.27 | 694.66 | 2.79x |
| Apple M2 Ultra (+fp16+dotprod) | TinyLlama-1.1B-Chat-v1.0.Q2_K | 411.41 MiB | tg16 | 108.34 | 105.45 | 0.97x |

The gains are also enormous on Raspberry Pi. Having 2x to 3x better is huge. I've gotten F16 to go as fast as 80 tok/sec (not sure why it doesn't anymore, could potentially be due to cooling). However I'm noticing that prediction is slowing down a bit on RPI5. Did you do anything to change that? Once again, it could be cooling. If you have any ideas, send me a follow-up change. With tinyBLAS in many cases it'll punt control back to GGML when `n=1`. The special codepaths should only run when they add value.

| cpu_info | model_filename | size | test | t/s before | t/s after | t/s speedup |
| ---: | ---: | ---: | ---: | ---: | ---: | ---: |
| +fp16+dotprod |             TinyLlama-1.1B-Chat-v1.0.F16 |   2.05 GiB |         pp512 |           66.53 |           66.53 | 1.00x |
| +fp16+dotprod |             TinyLlama-1.1B-Chat-v1.0.F16 |   2.05 GiB |          tg16 |            4.26 |            4.26 | 1.00x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q8_0 | 1.09 GiB | pp512 | 44.92 | 55.41 | 1.23x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q8_0 | 1.09 GiB | tg16 | 8.38 | 7.90 | 0.94x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q6_K | 860.86 MiB | pp512 | 18.20 | 37.59 | 2.07x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q6_K | 860.86 MiB | tg16 | 11.48 | 9.66 | 0.84x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q5_K_M | 745.11 MiB | pp512 | 19.38 | 41.25 | 2.13x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q5_K_M | 745.11 MiB | tg16 | 13.41 | 10.22 | 0.76x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q5_0 | 729.84 MiB | pp512 | 17.64 | 46.45 | 2.63x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q5_0 | 729.84 MiB | tg16 | 11.83 | 11.12 | 0.94x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q2_K | 411.41 MiB | pp512 | 18.80 | 44.74 | 2.38x |
| +fp16+dotprod | TinyLlama-1.1B-Chat-v1.0.Q2_K | 411.41 MiB | tg16 | 14.54 | 14.79 | 1.02x |